doc: extensions: options_from_kconfig: move to single file module
Make extension a single file module.
